RD-1526 Extract marker tooltip to a separate file
