+++ widgets/common/src/deploymentsView/map/marker/index.tsx
@@ -1,9 +1,8 @@
-import { camelCase } from 'lodash';
 import type { FunctionComponent } from 'react';
 
-import { Deployment, DeploymentStatus } from '../types';
-import { DeploymentSitePair, mapT } from './common';
-import { DeploymentStatusIcon } from '../StatusIcon';
+import { DeploymentStatus } from '../../types';
+import type { DeploymentSitePair } from '../common';
+import DeploymentSiteTooltip from './tooltip';
 
 const deploymentStatusToIconColorMapping: Record<DeploymentStatus, Stage.Common.MarkerIconColor> = {
     [DeploymentStatus.Good]: 'blue',
@@ -67,31 +66,5 @@ const BareDeploymentSiteMarker: FunctionComponent<{
         <Marker icon={icon} position={position} riseOnHover onclick={onClick}>
             {children}
         </Marker>
-    );
-};
-
-const tooltipStatusT = (suffix: string) => mapT(`tooltip.status.${suffix}`);
-const markerIconHeight = 41;
-const tooltipOffset = L.point(0, -markerIconHeight);
-const DeploymentSiteTooltip: FunctionComponent<{ deployment: Deployment; environmentTypeVisible: boolean }> = ({
-    deployment,
-    environmentTypeVisible
-}) => {
-    const {
-        Leaflet: { Tooltip },
-        Header
-    } = Stage.Basic;
-
-    return (
-        <Tooltip direction="top" offset={tooltipOffset}>
-            <Header as="h4">{deployment.id}</Header>
-            <div>{deployment.blueprint_id}</div>
-            <div>{deployment.site_name}</div>
-            {environmentTypeVisible && <div>{deployment.environment_type}</div>}
-            <div>
-                <DeploymentStatusIcon status={deployment.deployment_status} />
-                {tooltipStatusT(camelCase(deployment.deployment_status))}
-            </div>
-        </Tooltip>
     );
 };
